+++ dce-service/src/main/java/com/dce/business/service/impl/award/AwardServiceImpl.java
@@ -25,7 +25,7 @@ public class AwardServiceImpl implements IAwardService {
 
 
 	
-	public void calcAward(Integer buyUserId, Integer buyQty, Long orderId) {
+	public void calcAward(Integer buyUserId, Integer buyQty, Integer orderId) {
 		Assert.notNull(buyUserId, "购买者用户ID不能为空");
 		Assert.notNull(buyQty, "购买数量不能为空");
 		Assert.notNull(orderId, "购买订单ID不能为空");
