+++ dce-service/src/main/java/com/dce/business/service/impl/award/BuyerUpgrade.java
@@ -49,7 +49,7 @@ public class BuyerUpgrade implements IAwardCalculator {
 	 * @return
 	 */
 	@Override
-	public void doAward(int buyUserId, int buyQty, Long orderId) {
+	public void doAward(int buyUserId, int buyQty, Integer orderId) {
 		
 		//获取用户信息
 		UserDo  buyer = userService.getUser(buyUserId);
